requirements.txt
# Core dependencies
python-dotenv==1.0.0
openai==1.3.0
numpy==1.24.3

# Document processing
PyPDF2==3.0.1
python-docx==0.8.11

# NLP and embeddings
sentence-transformers==2.2.2
transformers>=4.36.0
huggingface-hub>=0.19.4
torch>=2.1.0
tokenizers>=0.15.0

# Vector storage and search
faiss-cpu==1.7.4

# Text processing
langchain==0.0.330
tiktoken==0.5.1